refactor: Moves code to sse and protobuf modules
